requirements.txt
streamlit
pyttsx3
pandas
Pillow
requests
openai>=1.0.0
pyobjc
tensorflow 
torch 
torchvision 
transformers 
opencv-python-headless
python-dotenv
langchain
pypdf 
langchain-community 
langchain-openai 
sentence-transformers 
chroma-hnswlib 
chromadb 
rank-bm25 
beautifulsoup4 
wikipedia 
langchainhub 
tabulate 
langchain-experimental
pypdf 
langchain_chroma
